% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{frequency_tables}
\alias{frequency_tables}
\title{Calculate the frequency table of all \code{data}'s variables.}
\usage{
frequency_tables(data)
}
\arguments{
\item{data}{A dataframe.}
}
\value{
A list containing the frequency table from all \code{data}'s variables.
}
\description{
Calculate the frequency table of all \code{data}'s variables.
}
\examples{
data(mtcars)
frequency_tables(mtcars)
}
