% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_centroids.R
\name{initial_traditional}
\alias{initial_traditional}
\title{Initializes the clusters using the traditional approach.}
\usage{
initial_traditional(data, k)
}
\arguments{
\item{data}{A dataframe.}

\item{k}{A integer number.}
}
\value{
A dataframe containing \code{k} rows randomly selected from \code{data}.
}
\description{
Initializes the clusters using the traditional approach.
}
\examples{
data(mtcars)
k <- 3
initial_traditional(mtcars, k)
}
