% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_centroids.R
\name{sample_weighted_category}
\alias{sample_weighted_category}
\title{Selects, using weighted sampling, a category from the table.}
\usage{
sample_weighted_category(table)
}
\arguments{
\item{table}{A prop.table.}
}
\value{
A vector of unit size containing the selected randomly weighted category.
}
\description{
Selects, using weighted sampling, a category from the table.
}
\examples{
data(mtcars)
table <- prop.table(table(mtcars$carb))
sample_weighted_category(table)
}
